dh-virtualenv forces executable to 'python2.7'
This doesn't really ensure new compatibility other than on an odd trusty install.
It was *intended* to add compatibility for Lucid packaging, but there are features of
upstart that are not in the Lucid version (setuid and setgid are in >=1.4, while
Lucid appears to run 0.9). So if you overrode the default Python with an old one,
this commit should save your bacon - otherwise it should be a NoOp.